Feat : 좋아요 확인 기능 시설 상세페이지 클릭 시 liked를 통해 좋아요를 확인할 수 있습니다

+++ src/main/java/backend/dev/likes/service/LikeService.java
@@ -44,12 +44,16 @@ public class LikeService {
         String userId = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = userRepository.findById(userId).orElseThrow(() -> new PublicPlusCustomException(ErrorCode.NOT_FOUND_USER));
         FacilityDetails facilityDetails = facilityDetailsRepository.getReferenceById(facilityId);
-        try{
-
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        if (likeRepository.existsByUserAndFacility(user,facilityDetails)) {
+            likeRepository.deleteByUserAndFacility(user,facilityDetails);
         }
-        likeRepository.deleteByUserAndFacility(user,facilityDetails);
         facilityDetails.changeLikes(facilityDetails.getLikes()-1);
+    }
+    public boolean getLikeStatus(String facilityId) {
+        String userId = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userRepository.findById(userId).orElseThrow(() -> new PublicPlusCustomException(ErrorCode.NOT_FOUND_USER));
+        FacilityDetails facilityDetails = facilityDetailsRepository.getReferenceById(facilityId);
+
+        return likeRepository.existsByUserAndFacility(user,facilityDetails);
     }
 }
